Revert "Merge pull request #5 from y-yagi/refactor"
This reverts commit 5ce68ffb7a1e0ebea12e3cdf9f4602f580b54d3b, reversing
changes made to b6fc6f2e1f2c9ef3bb2cb8ee6cadcc1f5c9659e9.

Reason: The current implementation has the issue of finding
Well-Known Types. This issue would be solved by #6.
But #6 requires the new version of google-protobuf.
