+++ grpc_reflection.gemspec
@@ -30,5 +30,4 @@ Gem::Specification.new do |spec|
   spec.require_paths = ["lib"]
 
   spec.add_dependency "grpc"
-  spec.add_dependency "google-protobuf", ">= 4.30.0.rc1"
 end
