+++ lib/grpc_reflection.rb
@@ -2,7 +2,8 @@
 
 require_relative "grpc_reflection/version"
 require_relative "grpc_reflection/file_descriptor_decorator"
-require_relative "grpc_reflection/file_descriptor_manager"
+require_relative "grpc_reflection/file_descriptor_decorator_manager"
+require_relative "grpc_reflection/protobuf_ext"
 require_relative "grpc_reflection/server"
 require_relative "grpc_reflection/server_alpha"
