+++ test/test_grpc_reflection.rb
@@ -65,7 +65,6 @@ class TestGrpcReflection < Minitest::Test
     response = stub.server_reflection_info([request]).first
 
     assert response.file_descriptor_response
-
     parsed = Google::Protobuf::FileDescriptorProto.decode(response.file_descriptor_response.file_descriptor_proto.first)
     assert_equal "test/protos/helloworld.proto", parsed.name
   end
